ai_2/requirements.txt
streamlit==1.36.0 
PyPDF2==3.0.1 
python-dotenv==1.0.1 
langchain==0.3.0 
langchain-community==0.3.0 
langchain-google-genai==2.1.5 
faiss-cpu==1.7.4 
google-api-core==2.25.1 
numpy==1.26.4